fix replica removing tests
