+++ plenum/test/replica_removing/test_replica_removing_with_backup_degraded_acc_quorum.py
@@ -1,0 +1,24 @@
+import pytest
+
+from plenum.test.replica_removing.helper import do_test_replica_removing_with_backup_degraded, replica_removing
+
+
+@pytest.fixture(scope="module")
+def tconf(tconf):
+    with replica_removing(tconf, acc_monitor_enabled=False, replica_remove_stratgey="quorum"):
+        yield tconf
+
+
+def test_replica_removing_with_backup_degraded(looper,
+                                               txnPoolNodeSet,
+                                               sdk_pool_handle,
+                                               sdk_wallet_client,
+                                               sdk_wallet_steward,
+                                               tconf,
+                                               tdir,
+                                               allPluginsPath):
+    do_test_replica_removing_with_backup_degraded(looper,
+                                                  txnPoolNodeSet,
+                                                  sdk_pool_handle,
+                                                  sdk_wallet_client,
+                                                  tconf)
